replace EBean with Spring Repositories
Let's don't overcomplicate things for now. I definitely think
it could be beneficial if we seperate the persistence
from the domain but at the moment I don't see any benefit
doing so, because PROX is most likely CRUD-based. We can
rethink the decision at a later point. For now it is more
important to get the domain- and application logic running